+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [0-based versioning](https://0ver.org/).
 
-## [Unreleased]
+## v0.14.3 - 18-October-2025
 
 ### Changed